requirements.txt
AfricastalkingGateway==1.8
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.4
django-cors-headers==2.1.0
django-filter==1.0.4
django-rest-swagger==2.1.2
django-tables2==1.10.0
djangorestframework==3.8.2
drf-openapi==1.3.0
gunicorn==19.0.0
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
lipisha==0.2.4
Markdown==2.6.8
MarkupSafe==1.0
openapi-codec==1.3.2
passlib==1.7.1
phonenumbers==8.9.1
psycopg2==2.7.3
pytz==2017.2
requests==2.18.3
simplejson==3.14.0
stripe==1.62.0
uritemplate==3.0.0
urllib3==1.22
